Split slider into slider (for single values) and slider-range (for ranges)
Unfortunately both have slightly different HTML, which used to be handled
by compiling different templates, but that doesn't work inside transcluded
scopes. brianmcd committed a workaround, which was to use a link function
and no templates, but this was by his admission a hack which only worked
with range sliders. This commit duplicates the directive into 'slider', a
single value slider, and 'slider-range', a range slider.

The code for each is surprisingly different, so there's not a huge amount
of duplication, though I could definitely extract some common
functionality out.